Added files via upload
Function for the briefing/mission notes that only admins will see. Just a simple copy and paste of a combination of line 1 to 28 of briefing.sqf and the whole file of f_brieing_admin.sqf of the F3 template. The part of the f_briefing_admin.sqf was pasted after line 24 of briefing.sqf.